+++ CHANGELOG.md
@@ -25,7 +25,7 @@
   `.git/*`, `.gitignore` and `.github/*` files.
 - The `gleam new` command gains the `--skip-github` flag to skip creation of
   `.github/*` files.
-
+- Added support for deno runtime for JavaScript target.
 
 ## v0.25.3 - 2022-12-16